refactor: move `n.consistency_check` subfunctions to own module
